Remove setProfileForProperty in favor of selectProperty and setProfileID.
